refactor config and identity related code to their own files
